refactor processTextureBuffer into loadTextureFile
